refactor: interceptor 로직 정리
- 로그인 처리로 착각해서 basic에는 누락한 인증처리 추가
- 기존 멤버인지 체크하는 로직을 인증요청 처리시점에 처리되도록 이동
