+++ src/main/java/nextstep/security/authentication/BasicAuthenticationInterceptor.java
@@ -10,25 +10,23 @@ import javax.servlet.http.HttpServletResponse;
 @Component
 public
 class BasicAuthenticationInterceptor implements HandlerInterceptor {
-
-    private final UserAuthenticationService userAuthenticationService;
-
     private final BasicAuthenticationDecoder basicAuthenticationDecoder;
+    private final AuthenticationProvider authenticationProvider;
 
-    public BasicAuthenticationInterceptor(UserAuthenticationService userAuthenticationService, BasicAuthenticationDecoder basicAuthenticationDecoder) {
-        this.userAuthenticationService = userAuthenticationService;
+    public BasicAuthenticationInterceptor(BasicAuthenticationDecoder basicAuthenticationDecoder, AuthenticationProvider authenticationProvider) {
         this.basicAuthenticationDecoder = basicAuthenticationDecoder;
+        this.authenticationProvider = authenticationProvider;
     }
 
     @Override
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) {
+        authenticationProvider.clearAuthentication();
 
         BasicAuthenticationToken basicAuthentication = basicAuthenticationDecoder.decode(request);
+        String principal = basicAuthentication.getPrincipal().toString();
+        String credentials = basicAuthentication.getCredentials().toString();
 
-        userAuthenticationService.validateMember(
-                basicAuthentication.getPrincipal().toString(),
-                basicAuthentication.getCredentials().toString()
-        );
+        authenticationProvider.doAuthentication(new BasicAuthenticationToken(principal, credentials));
 
         return true;
     }
