+++ src/main/java/nextstep/security/authentication/FormAuthenticationInterceptor.java
@@ -1,9 +1,7 @@
 package nextstep.security.authentication;
 
-import nextstep.security.support.SecurityContextHolder;
 import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.HandlerInterceptor;
-import org.springframework.web.servlet.ModelAndView;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -13,25 +11,21 @@ import java.util.Map;
 public
 class FormAuthenticationInterceptor implements HandlerInterceptor {
 
-    private final UserAuthenticationService userAuthenticationService;
-    private final FormAuthenticationProvider formAuthenticationProvider;
+    private final AuthenticationProvider authenticationProvider;
 
-    FormAuthenticationInterceptor(UserAuthenticationService userAuthenticationService, FormAuthenticationProvider formAuthenticationProvider) {
-        this.userAuthenticationService = userAuthenticationService;
-        this.formAuthenticationProvider = formAuthenticationProvider;
+    FormAuthenticationInterceptor(AuthenticationProvider authenticationProvider) {
+        this.authenticationProvider = authenticationProvider;
     }
 
     @Override
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
-        SecurityContextHolder.setContext(SecurityContextHolder.createEmptyContext());
+        authenticationProvider.clearAuthentication();
 
         Map<String, String[]> paramMap = request.getParameterMap();
-        String email = paramMap.get("username")[0];
-        String password = paramMap.get("password")[0];
+        String principal = paramMap.get("username")[0];
+        String credentials = paramMap.get("password")[0];
 
-        userAuthenticationService.validateMember(email, password);
-
-        formAuthenticationProvider.doAuthentication(new FormAuthenticationToken(email, password));
+        authenticationProvider.doAuthentication(new FormAuthenticationToken(principal, credentials));
 
         return HandlerInterceptor.super.preHandle(request, response, handler);
     }
